requirements.txt
Flask
flask-cors
numpy
Pillow
tflite-runtime
